+++ packages/world/src/Module.sol
@@ -12,6 +12,8 @@ import { InstalledModules } from "./codegen/tables/InstalledModules.sol";
  * @dev Abstract contract that implements the ERC-165 supportsInterface function for IModule.
  */
 abstract contract Module is IModule, WorldContextConsumer {
+  address internal immutable __self = address(this);
+
   /**
    * @notice Checks if the given interfaceId is supported by this contract.
    * @dev Overrides the functionality from IERC165 and WorldContextConsumer to check for supported interfaces.
@@ -29,21 +31,21 @@ abstract contract Module is IModule, WorldContextConsumer {
 
   /**
    * @dev Check if a module with the given name and arguments is installed.
-   * @param moduleName The name of the module.
+   * @param moduleAddress The address of the module.
    * @param args The arguments for the module installation.
    * @return true if the module is installed, false otherwise.
    */
-  function isInstalled(bytes16 moduleName, bytes memory args) internal view returns (bool) {
-    return InstalledModules.get(moduleName, keccak256(args)) != address(0);
+  function isInstalled(address moduleAddress, bytes memory args) internal view returns (bool) {
+    return InstalledModules.get(moduleAddress, keccak256(args));
   }
 
   /**
    * @dev Revert if the module with the given name and arguments is already installed.
-   * @param moduleName The name of the module.
+   * @param moduleAddress The address of the module.
    * @param args The arguments for the module installation.
    */
-  function requireNotInstalled(bytes16 moduleName, bytes memory args) internal view {
-    if (isInstalled(moduleName, args)) {
+  function requireNotInstalled(address moduleAddress, bytes memory args) internal view {
+    if (isInstalled(moduleAddress, args)) {
       revert Module_AlreadyInstalled();
     }
   }