+++ packages/world/src/modules/core/implementations/BatchCallSystem.sol
@@ -2,10 +2,8 @@
 pragma solidity >=0.8.0;
 
 import { System } from "../../../System.sol";
-import { LimitedCallContext } from "../LimitedCallContext.sol";
 import { IBaseWorld } from "../../../codegen/interfaces/IBaseWorld.sol";
 import { revertWithBytes } from "../../../revertWithBytes.sol";
-
 import { SystemCallData, SystemCallFromData } from "../types.sol";
 import { LimitedCallContext } from "../LimitedCallContext.sol";