+++ packages/world/test/Factories.t.sol
@@ -9,7 +9,6 @@ import { WORLD_VERSION } from "../src/version.sol";
 import { World } from "../src/World.sol";
 import { ResourceId } from "../src/WorldResourceId.sol";
 import { CoreModule } from "../src/modules/core/CoreModule.sol";
-import { CORE_MODULE_NAME } from "../src/modules/core/constants.sol";
 import { Create2Factory } from "../src/Create2Factory.sol";
 import { WorldFactory } from "../src/WorldFactory.sol";
 import { IWorldFactory } from "../src/IWorldFactory.sol";
@@ -78,11 +77,8 @@ contract FactoriesTest is Test, GasReporter {
     // Set the store address manually
     StoreSwitch.setStoreAddress(calculatedAddress);
 
-    // Retrieve CoreModule address from InstalledModule table
-    address installedModule = InstalledModules.get(CORE_MODULE_NAME, keccak256(new bytes(0)));
-
     // Confirm correct Core is installed
-    assertEq(installedModule, address(coreModule));
+    assertTrue(InstalledModules.get(address(coreModule), keccak256(new bytes(0))));
 
     // Confirm worldCount (which is salt) has incremented
     assertEq(uint256(worldFactory.worldCount()), uint256(1));