+++ packages/react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 2.0.0-next.16
+
+### Patch Changes
+
+- Updated dependencies [c6c13f2e]
+- Updated dependencies [e6c03a87]
+- Updated dependencies [37c228c6]
+- Updated dependencies [1bf2e908]
+- Updated dependencies [37c228c6]
+- Updated dependencies [37c228c6]
+- Updated dependencies [7b28d32e]
+- Updated dependencies [9f8b84e7]
+- Updated dependencies [ad4ac445]
+- Updated dependencies [37c228c6]
+- Updated dependencies [37c228c6]
+- Updated dependencies [37c228c6]
+- Updated dependencies [3ac68ade]
+- Updated dependencies [37c228c6]
+- Updated dependencies [103f635e]
+  - @latticexyz/store@2.0.0-next.16
+  - @latticexyz/recs@2.0.0-next.16
+
 ## 2.0.0-next.15
 
 ### Patch Changes